requirements.txt
streamlit
pymongo
python-dotenv
scikit-learn
numpy